Update admin interfaces
