+++ src/admin/AddStudentDetails.java
@@ -1,0 +1,210 @@
+package admin;
+
+import database.Connect;
+import java.awt.HeadlessException;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import javax.swing.JOptionPane;
+
+/**
+ *
+ * @author rwothoromo
+ */
+public class AddStudentDetails extends javax.swing.JFrame {
+
+    Connection conn = null;
+
+    /**
+     * Class constructor.
+     */
+    public AddStudentDetails() {
+        initComponents();
+        conn = Connect.dbConnector();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jLabelInstructions = new javax.swing.JLabel();
+        jLabelCourseCode = new javax.swing.JLabel();
+        txtCourseCode = new javax.swing.JTextField();
+        jLabelYrOfEntry = new javax.swing.JLabel();
+        txtYrOfEntry = new javax.swing.JTextField();
+        btnUpload = new javax.swing.JButton();
+        btnBack = new javax.swing.JButton();
+        jLabelLogo = new javax.swing.JLabel();
+        jLabelSystemName = new javax.swing.JLabel();
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+        setTitle("Testimonial Generating System - Add Student Details");
+
+        jLabelInstructions.setText("Type the details as required to match the name of the file you will upload.");
+
+        jLabelCourseCode.setFont(new java.awt.Font("Tahoma", 0, 18)); // NOI18N
+        jLabelCourseCode.setText("Course Code :");
+
+        txtCourseCode.setFont(new java.awt.Font("Tahoma", 0, 18)); // NOI18N
+
+        jLabelYrOfEntry.setFont(new java.awt.Font("Tahoma", 0, 18)); // NOI18N
+        jLabelYrOfEntry.setText("Year of Entry :");
+
+        txtYrOfEntry.setFont(new java.awt.Font("Tahoma", 0, 18)); // NOI18N
+
+        btnUpload.setFont(new java.awt.Font("Tahoma", 0, 18)); // NOI18N
+        btnUpload.setText("Upload");
+        btnUpload.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnUploadActionPerformed(evt);
+            }
+        });
+
+        btnBack.setFont(new java.awt.Font("Tahoma", 0, 18)); // NOI18N
+        btnBack.setText("Back");
+        btnBack.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnBackActionPerformed(evt);
+            }
+        });
+
+        jLabelLogo.setIcon(new javax.swing.ImageIcon(getClass().getResource("/images/kyuLogo.png"))); // NOI18N
+
+        jLabelSystemName.setFont(new java.awt.Font("Calibri", 1, 36)); // NOI18N
+        jLabelSystemName.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabelSystemName.setText("TESTIMONIAL GENERATING SYSTEM");
+        jLabelSystemName.setToolTipText("TGS");
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addContainerGap(50, Short.MAX_VALUE)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jLabelSystemName, javax.swing.GroupLayout.PREFERRED_SIZE, 755, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabelLogo, javax.swing.GroupLayout.PREFERRED_SIZE, 660, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabelInstructions)
+                    .addGroup(layout.createSequentialGroup()
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(jLabelYrOfEntry)
+                            .addComponent(jLabelCourseCode))
+                        .addGap(50, 50, 50)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                            .addGroup(layout.createSequentialGroup()
+                                .addComponent(btnUpload, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                                .addComponent(btnBack, javax.swing.GroupLayout.DEFAULT_SIZE, 200, Short.MAX_VALUE))
+                            .addComponent(txtYrOfEntry, javax.swing.GroupLayout.Alignment.TRAILING)
+                            .addComponent(txtCourseCode, javax.swing.GroupLayout.Alignment.TRAILING))))
+                .addContainerGap(50, Short.MAX_VALUE))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addContainerGap(50, Short.MAX_VALUE)
+                .addComponent(jLabelLogo)
+                .addGap(18, 18, 18)
+                .addComponent(jLabelSystemName, javax.swing.GroupLayout.PREFERRED_SIZE, 32, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(18, 18, 18)
+                .addComponent(jLabelInstructions)
+                .addGap(18, 18, 18)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(txtCourseCode, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabelCourseCode))
+                .addGap(18, 18, 18)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(txtYrOfEntry, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabelYrOfEntry))
+                .addGap(18, 18, 18)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(btnUpload, javax.swing.GroupLayout.PREFERRED_SIZE, 57, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(btnBack, javax.swing.GroupLayout.PREFERRED_SIZE, 57, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap(50, Short.MAX_VALUE))
+        );
+
+        pack();
+        setLocationRelativeTo(null);
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void btnBackActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnBackActionPerformed
+        dispose();
+        Dashboard SU = new Dashboard();
+        SU.setVisible(true);
+    }//GEN-LAST:event_btnBackActionPerformed
+
+    private void btnUploadActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnUploadActionPerformed
+        String Coursecode = txtCourseCode.getText();
+        String EntryYr = txtYrOfEntry.getText();
+        String db = Coursecode + EntryYr;
+        String sqlUpload = "LOAD DATA INFILE 'd:/TGS datafiles/" + db + ".csv'"
+                + "INTO TABLE " + db + " FIELDS TERMINATED BY ','"
+                + " ENCLOSED BY '\"' LINES TERMINATED BY '\\n' IGNORE 1 ROWS;";
+        //ignores the title row
+        String checkDB = "SELECT * FROM " + db;
+        try {
+            PreparedStatement ps = conn.prepareStatement(checkDB);
+            PreparedStatement ps1 = conn.prepareStatement(sqlUpload);
+            ResultSet rs = ps.executeQuery();
+            if (rs.next()) {
+                ps1.executeQuery(sqlUpload);
+                JOptionPane.showMessageDialog(null, db + ".csv has been uploaded!");
+            }
+        } catch (HeadlessException | SQLException e) {
+            JOptionPane.showMessageDialog(null, e);
+        }
+    }//GEN-LAST:event_btnUploadActionPerformed
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String args[]) {
+        /* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(AddStudentDetails.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(AddStudentDetails.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(AddStudentDetails.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(AddStudentDetails.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+
+        /* Create and display the form */
+        java.awt.EventQueue.invokeLater(() -> {
+            new AddStudentDetails().setVisible(true);
+        });
+    }
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton btnBack;
+    private javax.swing.JButton btnUpload;
+    private javax.swing.JLabel jLabelCourseCode;
+    private javax.swing.JLabel jLabelInstructions;
+    private javax.swing.JLabel jLabelLogo;
+    private javax.swing.JLabel jLabelSystemName;
+    private javax.swing.JLabel jLabelYrOfEntry;
+    private javax.swing.JTextField txtCourseCode;
+    private javax.swing.JTextField txtYrOfEntry;
+    // End of variables declaration//GEN-END:variables
+}
