+++ src/admin/Dashboard.form
@@ -44,15 +44,10 @@
                           </Group>
                           <Component id="btnSignOut" min="-2" pref="211" max="-2" attributes="0"/>
                       </Group>
+                      <EmptySpace type="separate" max="-2" attributes="0"/>
                       <Group type="103" groupAlignment="0" attributes="0">
-                          <Group type="102" attributes="0">
-                              <EmptySpace type="separate" max="-2" attributes="0"/>
-                              <Component id="btnProcessTestimonial" min="-2" pref="211" max="-2" attributes="0"/>
-                          </Group>
-                          <Group type="102" alignment="0" attributes="0">
-                              <EmptySpace type="separate" max="-2" attributes="0"/>
-                              <Component id="btnChangePassword" min="-2" pref="211" max="-2" attributes="0"/>
-                          </Group>
+                          <Component id="btnProcessTestimonial" min="-2" pref="211" max="-2" attributes="0"/>
+                          <Component id="btnChangePassword" alignment="0" min="-2" pref="211" max="-2" attributes="0"/>
                       </Group>
                   </Group>
               </Group>
